Removing memoizing code and keeping backfilling code.
This commit has all of the notes and commented out code related to
memoizing function calls. Leaving for historical paper trail of the
work.

I originally made the branch to try to enable memoizing of fillers, aka
assigning the filler to a variable and reusing that as the filler. I got
that working, then started trying to do backfilling. The combinatorial
logic of trying to make these things work together turned into a
shit-nest of logic.

I then ripped out all of the memoization code and instead implemented
backfilling, which ended up being way simpler, by modifying the filler
to take a function that returns a node, rather than an node.

This branch is called "memoize" but I'm leaving it as backfilling for
now. This commit has some of the memoization logic that might help
re-implement it.

Memoizing is probably good for other strategies, like assignTo, where
you're replacing an expression outright, so there probably isn't
anything to backfill. For texture2D() calls it probably doesn't make
sense, and instead it's up to the shader author to determine when to
call texture2D() as it's less likely there will be fully duplicative
calls there.